:label: [#445] Convert Literal component to TS
